requirements.txt
# VUTT Backend dependencies
GitPython>=3.1.40
